requirements.txt
matplotlib==3.10.1
numpy==2.2.4
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
PyYAML==6.0.2
rembg==2.0.64
scikit_learn==1.6.1
torch==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
ultralytics==8.3.73
